Support rust-nightly-2022-10-22 and Bump version to 1.1.6

+++ sgx_align_struct_attribute/src/layout.rs
@@ -124,7 +124,7 @@ fn check_align_req(size: usize, align_req: &[AlignReq]) -> bool {
 
 fn gen_alignmask(al: usize, a: usize, m: u64) -> i64 {
     if a > al {
-        gen_alignmask(al, (a >> 1) as usize, m | (m >> (a >> 1)))
+        gen_alignmask(al, a >> 1, m | (m >> (a >> 1)))
     } else {
         m as i64
     }
